Renamed python config from `TrackerAlignmentProducerForPCL_cff` to generic `AlignmentProducerAsAnalyzer_cff`.
