+++ Alignment/CommonAlignmentProducer/python/AlcaSiPixelAliHarvester_cff.py
@@ -10,7 +10,7 @@ SiPixelAliMilleFileExtractor = cms.EDAnalyzer("MillePedeFileExtractor",
     outputBinaryFile = cms.string('pedeBinary%04d.dat'))
 
 from Alignment.MillePedeAlignmentAlgorithm.MillePedeAlignmentAlgorithm_cfi import *
-from Alignment.CommonAlignmentProducer.TrackerAlignmentProducerForPCL_cff import AlignmentProducer
+from Alignment.CommonAlignmentProducer.AlignmentProducerAsAnalyzer_cff import AlignmentProducer
 SiPixelAliPedeAlignmentProducer = copy.deepcopy(AlignmentProducer)
 
 from Alignment.MillePedeAlignmentAlgorithm.MillePedeDQMModule_cff import *
